streamlit_app/requirements.txt
openai
panda
streamlit
Pillow
opencv-python
numpy
dlib